Update to new computer

+++ Nothing.py
@@ -1,76 +1,44 @@
-from Account import Account
+from accounts import Account
+
+def chart_menu():
+
+    print("")
+    print("Chart of Accounts Menu: ")
+    print("1 - View Current Chart of Accounts")
+    print("2 - Add an Account")
+    print("3 - Delete an Account")
+    print("4 - Update an Account Attribute")
+    print("5 - Clear Chart of Accounts")
+    print("6 - Return to Default Accounts")
+    print("7 - Exit")
+    print("")
 
 
-class ChartOfAccountsApp(object):
-    
-    def __init__(self):
-        
-        self.chart_of_accounts = [  # Default Chart
-            Account(acc_num=int(1000), acc_name="Cash", acc_type="Asset"),
-            Account(acc_num=int(2000), acc_name="Accounts Payable", acc_type="Liability"),
-            Account(acc_num=int(3000), acc_name="Retained Earnings", acc_type="Equity"),
-            Account(acc_num=int(4000), acc_name="Sales Revenue", acc_type="Income"),
-            Account(acc_num=int(5000), acc_name="Cost of Goods Sold", acc_type="Expense"),
-        ]
-    
-    def run(self):
-        
-        print("")
-        print("Chart of Accounts Menu: ")
-        print("1 - View Current Chart of Accounts")
-        print("2 - Add an Account")
-        print("3 - Delete an Account")
-        print("4 - Update an Account Attribute")
-        print("5 - Clear Chart of Accounts")
-        print("6 - Return to Default Accounts")
-        print("7 - Exit")
-        print("")
+user_input = "None"
 
-        while True:
+print("")
+print("Welcome to the chart of accounts!")
+print("What would you like to do?")
 
-            user_input = input("Please choose an option")
+while True:
+    chart_menu()
 
-            if user_input == "1":
-                self.option_one()
+    user_input = input("Please choose an option")
 
-            elif user_input == "2":
-                self.option_two()
-                
-            elif user_input == "3":
-                self.option_three()
-                
-            elif user_input == "4":
-                self.option_four()
-                
-            elif user_input == "5":
-                self.option_five()
-                
-            elif user_input == "6":
-                self.option_six()
-                
-            elif user_input == "7":
-                break
-            else:
-                self.chart_menu()
-
-    
-    def option_one(self):
-        
-        if self.chart_of_accounts == []:
+    if user_input == "1":
+        if list_of_accounts == []:
             print("")
             print("No Chart of Accounts Available!")
         else:
             print("---------------------------------------")
-            print("{} - {} - {} - {}".format("Account #", "Account Name", "Account Type", "Beginning Balance"))
+            print("{} - {} - {} - {}".format("Account #", "Account Name", "Account Type", "Account Balance"))
             print("")
-            for account in self.chart_of_accounts:
+            for account in list_of_accounts:
                 print("{} - {} - {} - ${}".format(account.acc_num, account.acc_name, account.acc_type,
                                                   account.acc_balance))
             print("---------------------------------------")
-            
-            
-    def option_two(self):
-        
+
+    elif user_input == "2":
         while True:
             while True:
                 account_num = 0
@@ -80,7 +48,7 @@ class ChartOfAccountsApp(object):
                     print("Account # cannot contain letters")
                     break
                 num_in_use = 0
-                for account in self.chart_of_accounts:
+                for account in list_of_accounts:
                     if account.acc_num == account_num:
                         print("Account # already in use. Try again")
                         num_in_use = 1
@@ -91,23 +59,19 @@ class ChartOfAccountsApp(object):
             account_name = input("Account Name")
             account_type = input("Account Type")
             i = 0
-            num_of_accounts = len(self.chart_of_accounts)
+            num_of_accounts = len(list_of_accounts)
 
             while num_of_accounts > i:
-                if account_num < self.chart_of_accounts[i].acc_num:
-                    self.chart_of_accounts.insert(i, (
-                        Account(acc_num=int(account_num), acc_name=account_name, acc_type=account_type)))
+                if account_num < list_of_accounts[i].acc_num:
+                    list_of_accounts.insert(i, (Account(acc_num=int(account_num), acc_name=account_name, acc_type=account_type)))
                     break
                 i += 1
             else:
-                self.chart_of_accounts.append(
-                    Account(acc_num=int(account_num), acc_name=account_name, acc_type=account_type))
+                list_of_accounts.append(Account(acc_num=int(account_num), acc_name=account_name, acc_type=account_type))
             add_more = input("Type \"Y\" to enter another account, otherwise click anything else")
             if add_more.upper() != "Y":
                 break
-                
-    def option_three(self):
-        
+    elif user_input == "3":
         removed = 0
         delete_num = 0
         try:
@@ -115,16 +79,14 @@ class ChartOfAccountsApp(object):
         except ValueError:
             print("Account # cannot contain letters")
         if delete_num != 0:
-            for account in self.chart_of_accounts:
+            for account in list_of_accounts:
                 if account.acc_num == delete_num:
-                    self.chart_of_accounts.remove(account)
+                    list_of_accounts.remove(account)
                     print("Account #{} Removed!".format(delete_num))
                     removed = 1
         if removed == 0 and delete_num != 0:
             print("Sorry, there is no account #{}".format(delete_num))
-            
-    def option_four(self):
-        
+    elif user_input == "4":
         updated = 0
         update_num = 0
         try:
@@ -132,13 +94,11 @@ class ChartOfAccountsApp(object):
         except ValueError:
             print("Account # cannot contain letters and must be a whole number greater than 0")
             updated = 1
-        for account in self.chart_of_accounts:
+        for account in list_of_accounts:
             if account.acc_num == update_num:
                 print("")
-                print(
-                    "{} - {} - {} - {}".format("Account #", "Account Name", "Account Type", "Account Balance"))
-                print("{} - {} - {} - ${}".format(account.acc_num, account.acc_name, account.acc_type,
-                                                  account.acc_balance))
+                print("{} - {} - {} - {}".format("Account #", "Account Name", "Account Type", "Account Balance"))
+                print("{} - {} - {} - ${}".format(account.acc_num, account.acc_name, account.acc_type, account.acc_balance))
                 print("")
                 print("What attribute would you like to update for Acc # {}".format(update_num))
                 print("1 - Account #")
@@ -160,7 +120,7 @@ class ChartOfAccountsApp(object):
                         print("Sorry, the account number must be greater than 0")
                         updated = 1
                         break
-                    for accounts in self.chart_of_accounts:
+                    for accounts in list_of_accounts:
                         if accounts.acc_num == new_acc_num:
                             duplicate += 1
                     if duplicate == 1:
@@ -169,9 +129,8 @@ class ChartOfAccountsApp(object):
                     else:
                         account.acc_num = new_acc_num
                         print("Account # Changed!")
-                        sorted_list = sorted(self.chart_of_accounts, key=lambda Accounts: Accounts.acc_num,
-                                             reverse=False)
-                        self.chart_of_accounts = sorted_list
+                        sorted_list = sorted(list_of_accounts, key=lambda Accounts: Accounts.acc_num, reverse=False)
+                        list_of_accounts = sorted_list
                         updated = 1
                 elif update == "2":
                     new_acc_name = input("Please enter new account name: ")
@@ -199,36 +158,23 @@ class ChartOfAccountsApp(object):
                     updated = 1
         if updated == 0:
             print("Not a valid acc #")
-            
-    def option_five(self):
-        
+    elif user_input == "5":
         delete_all = input("Are you sure you want to delete your chart of accounts? Type \"Y\" to proceed ")
         if delete_all.upper() == "Y":
-            self.chart_of_accounts = []
+            list_of_accounts = []
             print("Chart of Accounts Deleted!")
         else:
             print("Chart of accounts not deleted.")
-            
-    def option_six(self):
-        self.chart_of_accounts = [  # Default Chart
+    elif user_input == "6":
+        list_of_accounts = [  # Default Chart
             Account(acc_num=int(1000), acc_name="Cash", acc_type="Asset"),
             Account(acc_num=int(2000), acc_name="Accounts Payable", acc_type="Liability"),
             Account(acc_num=int(3000), acc_name="Retained Earnings", acc_type="Equity"),
             Account(acc_num=int(4000), acc_name="Sales Revenue", acc_type="Income"),
             Account(acc_num=int(5000), acc_name="Cost of Goods Sold", acc_type="Expense"),
         ]
-
         print("Default Chart of accounts has been restored.")
-
-    def chart_menu(self):
-
-        print("")
-        print("Chart of Accounts Menu: ")
-        print("1 - View Current Chart of Accounts")
-        print("2 - Add an Account")
-        print("3 - Delete an Account")
-        print("4 - Update an Account Attribute")
-        print("5 - Clear Chart of Accounts")
-        print("6 - Return to Default Accounts")
-        print("7 - Exit")
-        print("")
+    elif user_input == "7":
+        break
+    else:
+        print("Not an option. Here are the available options:")
